+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - New parameter `commaSeparatedValue` for the `labelHasAllowedValue` validator supporting labels with a comma separated values.
 
 ## [v1.1.0] - 2020-11-20
 ### Changed